Add support for GCC on Windows with MSYS2-UCRT64 - Support querying CPU information on Windows fully, including some currently unused information such as the efficiency class. - Support thread affinity using CPU sets on Windows. - Support querying memory use and time consumption on Windows. - Fix an issue arising from 32-bit `long` on Windows. - Support Windows’s paths using 16 bit characters by converting to UTF-8. - Support escaping both `char`- and `char8_t`-based string views, which is used for UTF8-converted paths on Windows. - Implement `HugePagesAllocator` with alignment, but without huge pages, as this requires additional - Extend the execution test with a test for the thread affinity code. - Set Windows 10 as the minimum supported OS version, as it is the first version that supports CPU sets. - Compile and link {{fmt}} as a static library, as using it as a shared library led to problems when executing the binary.

+++ test/json.cpp
@@ -18,6 +18,7 @@
 #include "thesauros/io.hpp"
 #include "thesauros/macropolis.hpp"
 #include "thesauros/test.hpp"
+#include "thesauros/types.hpp"
 
 namespace test = thes::test;
 
@@ -46,11 +47,12 @@ int main() {
                               thes::numeric_string(std::numeric_limits<f32>::max()).value()));
   THES_ASSERT(test::string_eq("2.2250738585072014e-308",
                               thes::numeric_string(std::numeric_limits<f64>::min()).value()));
-  THES_ASSERT(test::string_eq("-9223372036854775808",
-                              thes::numeric_string(std::numeric_limits<long>::lowest()).value()));
+  THES_ASSERT(
+    test::string_eq("-9223372036854775808",
+                    thes::numeric_string(std::numeric_limits<long long>::lowest()).value()));
   THES_ASSERT(
     test::string_eq("18446744073709551615",
-                    thes::numeric_string(std::numeric_limits<unsigned long>::max()).value()));
+                    thes::numeric_string(std::numeric_limits<unsigned long long>::max()).value()));
 
   // thes::escaped_string
 
@@ -63,10 +65,10 @@ int main() {
   THES_ASSERT(test::string_eq("3.4028235e+38", thes::json_print(std::numeric_limits<f32>::max())));
   THES_ASSERT(
     test::string_eq("2.2250738585072014e-308", thes::json_print(std::numeric_limits<f64>::min())));
-  THES_ASSERT(
-    test::string_eq("-9223372036854775808", thes::json_print(std::numeric_limits<long>::lowest())));
+  THES_ASSERT(test::string_eq("-9223372036854775808",
+                              thes::json_print(std::numeric_limits<long long>::lowest())));
   THES_ASSERT(test::string_eq("18446744073709551615",
-                              thes::json_print(std::numeric_limits<unsigned long>::max())));
+                              thes::json_print(std::numeric_limits<unsigned long long>::max())));
 
   THES_ASSERT(test::string_eq("\"abc\\\"\\r\\n\"", thes::json_print("abc\"\r\n")));
   THES_ASSERT(test::string_eq("\"\\u0007\"", thes::json_print("\a")));
